+++ client/src/components/worldmap/biomes/TropicalRainforestBiome.tsx
@@ -2,7 +2,7 @@ import { useGLTF } from "@react-three/drei";
 import { getUIPositionFromColRow } from "../../../utils/utils";
 import * as THREE from "three";
 import { useMemo } from "react";
-import { Hexagon } from "../HexGrid";
+import { Hexagon } from "src/types";
 import { GLTF } from "three-stdlib";
 
 type GLTFResult = GLTF & {
