+++ client/src/components/worldmap/hexagon/utils.tsx
@@ -1,0 +1,132 @@
+import { Color } from "three";
+import { Position, neighborOffsetsEven, neighborOffsetsOdd } from "@bibliothecadao/eternum";
+import { Hexagon } from "../../../types";
+
+const matrix = new Matrix4();
+const positions = new Vector3();
+import { InstancedMesh, Matrix4, Vector3 } from "three";
+
+export const findShortestPathDFS = (startPos: Position, endPos: Position, hexData: Hexagon[], maxHex: number) => {
+  const stack = [startPos];
+  const visited = new Set<Position>();
+  const path = new Map<string, Position>();
+  let count = 0;
+
+  const posKey = (pos: Position) => `${pos.x},${pos.y}`; // Create a unique string key for each position
+
+  while (stack.length > 0 && count <= maxHex) {
+    const current = stack.pop() as Position; // Use pop to take from the stack
+    if (current.x === endPos.x && current.y === endPos.y) {
+      const result = [current];
+      let next = path.get(posKey(current));
+      while (next) {
+        result.push(next);
+        next = path.get(posKey(next));
+      }
+      return result.reverse();
+    }
+
+    if (!visited.has(current)) {
+      visited.add(current);
+      const neighbors = getNeighbors(current, hexData);
+      count++;
+      for (const neighbor of neighbors) {
+        if (!visited.has(neighbor)) {
+          path.set(posKey(neighbor), current);
+          stack.push(neighbor); // Push neighbors onto the stack
+        }
+      }
+    }
+  }
+
+  return [];
+};
+
+export const isNeighbor = (pos1: Position, pos2: Position) => {
+  const neighborOffsets = pos1.y % 2 === 0 ? neighborOffsetsEven : neighborOffsetsOdd;
+  for (const { i, j } of neighborOffsets) {
+    if (pos1.x + i === pos2.x && pos1.y + j === pos2.y) {
+      return true;
+    }
+  }
+  return false;
+};
+
+export const getNeighbors = (pos: Position, hexData: Hexagon[]) => {
+  const neighborOffsets = pos.y % 2 === 0 ? neighborOffsetsEven : neighborOffsetsOdd;
+
+  return neighborOffsets
+    .map((offset) => {
+      const col = pos.x + offset.i;
+      const row = pos.y + offset.j;
+      const hex = hexData.find((h) => h.col === col && h.row === row);
+      return hex ? { x: hex.col, y: hex.row } : null;
+    })
+    .filter(Boolean) as Position[];
+};
+
+export const getGrayscaleColor = (color: Color) => {
+  const luminance = 0.299 * color.r + 0.587 * color.g + 0.114 * color.b;
+  return luminance;
+};
+
+export const getPositionsAtIndex = (mesh: InstancedMesh<any, any>, index: number) => {
+  if (!mesh || !mesh.isInstancedMesh) {
+    console.error("The provided mesh is not an InstancedMesh.");
+    return null;
+  }
+
+  mesh.getMatrixAt(index, matrix);
+  positions.setFromMatrixPosition(matrix);
+
+  return positions;
+};
+
+export const findShortestPathBFS = (
+  startPos: Position,
+  endPos: Position,
+  hexData: Hexagon[],
+  exploredHexes: Map<number, Set<number>>,
+  maxHex: number,
+) => {
+  const queue: { position: Position; distance: number }[] = [{ position: startPos, distance: 0 }];
+  const visited = new Set<string>();
+  const path = new Map<string, Position>();
+
+  const posKey = (pos: Position) => `${pos.x},${pos.y}`;
+
+  while (queue.length > 0) {
+    const { position: current, distance } = queue.shift()!;
+    if (current.x === endPos.x && current.y === endPos.y) {
+      // Reconstruct the path upon reaching the end position
+      let temp = current;
+      const result = [];
+      while (temp) {
+        result.unshift(temp); // Add to the beginning of the result array
+        //@ts-ignore:
+        temp = path.get(posKey(temp)); // Move backwards through the path
+      }
+      return result;
+    }
+
+    if (distance > maxHex) {
+      break; // Stop processing if the current distance exceeds maxHex
+    }
+
+    const currentKey = posKey(current);
+    if (!visited.has(currentKey)) {
+      visited.add(currentKey);
+      const neighbors = getNeighbors(current, hexData); // Assuming getNeighbors is defined elsewhere
+      for (const neighbor of neighbors) {
+        const neighborKey = posKey(neighbor);
+        const isExplored = exploredHexes.get(neighbor.x - 2147483647)?.has(neighbor.y - 2147483647);
+        if (!visited.has(neighborKey) && !queue.some((e) => posKey(e.position) === neighborKey) && isExplored) {
+          path.set(neighborKey, current); // Map each neighbor back to the current position
+          queue.push({ position: neighbor, distance: distance + 1 });
+        }
+      }
+    }
+  }
+
+  return []; // Return empty array if no path is found within maxHex distance
+};
