--- client/src/graphql/Labor/getRealmLabor.graphql
@@ -1,17 +1,0 @@
-query getRealmLabor($realmEntityId: String!) {
-  entities(keys: [$realmEntityId], first: 100) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on Labor {
-            last_harvest
-            balance
-            multiplier
-          }
-        }
-      }
-    }
-  }
-}
