--- client/src/graphql/Labor/getRealmResources.graphql
@@ -1,15 +1,0 @@
-query getRealmResources($realmEntityId: String!) {
-  entities(keys: [$realmEntityId], first: 50) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on Resource {
-            balance
-          }
-        }
-      }
-    }
-  }
-}
