--- client/src/graphql/Realm/getRealm.graphql
@@ -1,31 +1,0 @@
-query getRealm($realmEntityId: String!) {
-  entities(keys: [$realmEntityId], first: 100) {
-    edges {
-      node {
-        keys
-        id
-        components {
-          __typename
-          ... on Realm {
-            realm_id
-            cities
-            rivers
-            wonder
-            harbors
-            regions
-            resource_types_count
-            resource_types_packed
-            order
-          }
-          ... on Position {
-            x
-            y
-          }
-          ... on Owner {
-            address
-          }
-        }
-      }
-    }
-  }
-}
