--- client/src/graphql/Trade/Trade.fragment.graphql
@@ -1,10 +1,0 @@
-fragment TradeFragment on Trade {
-  maker_id
-  taker_id
-  maker_order_id
-  taker_order_id
-  expires_at
-  claimed_by_maker
-  claimed_by_taker
-  taker_needs_caravan
-}
