--- client/src/graphql/Trade/getCounterpartyOrderId.graphql
@@ -1,18 +1,0 @@
-query getCounterpartyOrderId($orderId: String!) {
-  makerTradeComponents: tradeComponents(where: { maker_order_id: $orderId }) {
-    edges {
-      node {
-        __typename
-        ...TradeFragment
-      }
-    }
-  }
-  takerTradeComponents: tradeComponents(where: { taker_order_id: $orderId }) {
-    edges {
-      node {
-        __typename
-        ...TradeFragment
-      }
-    }
-  }
-}
