--- client/src/graphql/Trade/getIncomingOrderInfo.graphql
@@ -1,43 +1,0 @@
-query getIncomingOrderInfo($orderId: String!, $counterPartyOrderId: String!) {
-  ## counterparty
-  resources: entities(keys: [$counterPartyOrderId]) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on OrderResource {
-            resource_type
-            balance
-          }
-          ... on FungibleEntities {
-            key
-            count
-          }
-          ... on ArrivalTime {
-            arrives_at
-          }
-          ... on Position {
-            x
-            y
-          }
-        }
-      }
-    }
-  }
-  # my offer
-  origin: entities(keys: [$orderId]) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on Position {
-            x
-            y
-          }
-        }
-      }
-    }
-  }
-}
