--- client/src/graphql/Trade/getIncomingOrders.graphql
@@ -1,28 +1,0 @@
-query getIncomingOrders($realmEntityId: String!) {
-  makerTradeComponents: tradeComponents(where: { maker_id: $realmEntityId, claimed_by_maker: 0 }) {
-    edges {
-      node {
-        entity {
-          keys
-          components {
-            __typename
-            ...TradeFragment
-          }
-        }
-      }
-    }
-  }
-  takerTradeComponents: tradeComponents(where: { taker_id: $realmEntityId, claimed_by_taker: 0 }) {
-    edges {
-      node {
-        entity {
-          keys
-          components {
-            __typename
-            ...TradeFragment
-          }
-        }
-      }
-    }
-  }
-}
