--- client/src/graphql/Trade/getMyOffers.graphql
@@ -1,18 +1,0 @@
-query getMyOffers($makerId: String!) {
-  tradeComponents(where: { maker_id: $makerId, claimed_by_maker: 0, claimed_by_taker: 0 }, last: 30) {
-    edges {
-      node {
-        entity {
-          keys
-          components {
-            __typename
-            ... on Status {
-              value
-            }
-            ...TradeFragment
-          }
-        }
-      }
-    }
-  }
-}
