--- client/src/graphql/Trade/getTradeResources.graphql
@@ -1,38 +1,0 @@
-query getTradeResources($makerOrderId: String!, $takerOrderId: String!) {
-  resourcesGive: entities(keys: [$makerOrderId], last: 100) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on OrderResource {
-            resource_type
-            balance
-          }
-          ... on FungibleEntities {
-            key
-            count
-          }
-        }
-      }
-    }
-  }
-  resourcesGet: entities(keys: [$takerOrderId], last: 100) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on OrderResource {
-            resource_type
-            balance
-          }
-          ... on FungibleEntities {
-            key
-            count
-          }
-        }
-      }
-    }
-  }
-}
