--- client/src/graphql/Travel/getCaravanInfo.graphql
@@ -1,72 +1,0 @@
-query getCaravanInfo($caravanId: String!, $orderId: String!, $counterpartyOrderId: String!) {
-  caravan: entities(keys: [$caravanId], first: 100) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on ArrivalTime {
-            arrives_at
-          }
-          ... on Movable {
-            blocked
-            sec_per_km
-          }
-          ... on Capacity {
-            weight_gram
-          }
-        }
-      }
-    }
-  }
-  destination: entities(keys: [$orderId]) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on Position {
-            x
-            y
-          }
-        }
-      }
-    }
-  }
-  resourcesGive: entities(keys: [$orderId]) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on OrderResource {
-            resource_type
-            balance
-          }
-          ... on FungibleEntities {
-            key
-            count
-          }
-        }
-      }
-    }
-  }
-  resourcesGet: entities(keys: [$counterpartyOrderId]) {
-    edges {
-      node {
-        keys
-        components {
-          __typename
-          ... on OrderResource {
-            resource_type
-            balance
-          }
-          ... on FungibleEntities {
-            key
-            count
-          }
-        }
-      }
-    }
-  }
-}
