--- client/src/graphql/Travel/getRealmCaravans.graphql
@@ -1,34 +1,0 @@
-query getRealmsCaravans($x: Int!, $y: Int!) {
-  positionComponents(where: { x: $x, y: $y }) {
-    edges {
-      node {
-        entity {
-          keys
-          components {
-            __typename
-            ... on OrderId {
-              id
-            }
-            ... on ArrivalTime {
-              arrives_at
-            }
-            ... on Movable {
-              blocked
-            }
-            ... on Capacity {
-              weight_gram
-            }
-            ... on Position {
-              x
-              y
-            }
-            ... on CaravanMembers {
-              key
-              count
-            }
-          }
-        }
-      }
-    }
-  }
-}
