+++ client/src/types/index.ts
@@ -1,0 +1,9 @@
+export interface Hexagon {
+  idx: number;
+  col: number;
+  row: number;
+  biome: string;
+  explored: boolean | undefined;
+  // address
+  exploredBy: bigint | undefined;
+}
